Improve performance for some use cases
The problem was this: The first object passed to merge was merged with itself. This is no big problem
when a lot of small objects are merged but if only two objects are merged, and the first one is much
bigger than the second one, this can lead to a drop in performance.